// src/store/main/analysis/type.ts
/*
 * @Author: Yico
 * @LastEditors: Yico
 * @Date: 2021-12-04 17:06:48
 * @LastEditTime: 2021-12-04 17:08:34
 * @Email: 2604482363@qq.com
 * @FilePath: \TEST_coder\src\store\main\analysis\type.ts
 * @Description:
 */
export interface IDashboardState {
  categoryGoodsCount: any[]
  categoryGoodsSale: any[]
  categoryGoodsFavor: any[]
  addressGoodsSale: any[]
}